+++ wxt.config.ts
@@ -7,8 +7,6 @@ export default defineConfig({
   manifest: {
     description:
       "Syntax highlighting and autosave for comments on GitHub (and other markdown-friendly websites).",
-    host_permissions: ["https://github.com/*"],
-    optional_host_permissions: ["https://*/*", "http://*/*"],
     icons: {
       16: "/icons/icon-16.png",
       48: "/icons/icon-48.png",